+++ five/intid/keyreference.py
@@ -1,22 +1,21 @@
-# -*- coding: utf-8 -*-
 from Acquisition import aq_base
 from Acquisition import aq_chain
 from Acquisition import IAcquirer
-from ZODB.interfaces import IConnection
-from ZPublisher.BaseRequest import RequestContainer
+from five.intid.site import get_root
+from five.intid.utils import aq_iter
 from persistent import IPersistent
+from ZODB.interfaces import IConnection
 from zope.component import adapter
 from zope.component.hooks import getSite
 from zope.interface import implementer
-from zope.keyreference.interfaces import NotYet
 from zope.keyreference.interfaces import IKeyReference
+from zope.keyreference.interfaces import NotYet
 from zope.keyreference.persistent import KeyReferenceToPersistent
-from five.intid.utils import aq_iter
-from five.intid.site import get_root
 from zope.lifecycleevent.interfaces import IObjectAddedEvent
+from ZPublisher.BaseRequest import RequestContainer
 
 import logging
-import six
+
 
 logger = logging.getLogger(__name__)
 
@@ -24,12 +23,12 @@ logger = logging.getLogger(__name__)
 @adapter(IPersistent)
 @implementer(IConnection)
 def connectionOfPersistent(obj):
-    """ zope2 cxn fetcher for wrapped items """
-    if not (IAcquirer.providedBy(obj) or hasattr(obj, '__parent__')):
-        return getattr(obj, '_p_jar', None)
+    """zope2 cxn fetcher for wrapped items"""
+    if not (IAcquirer.providedBy(obj) or hasattr(obj, "__parent__")):
+        return getattr(obj, "_p_jar", None)
 
     for parent in aq_iter(obj):
-        conn = getattr(parent, '_p_jar', None)
+        conn = getattr(parent, "_p_jar", None)
         if conn is not None:
             return conn
 
@@ -50,7 +49,7 @@ def traverse(base, path):
     raises: KeyError if not traversable this way
     """
     current = base
-    for cid in path.split('/'):
+    for cid in path.split("/"):
         if not cid:
             continue
         current = current[cid]
@@ -68,28 +67,28 @@ class KeyReferenceToPersistent(KeyReferenceToPersistent):
     @@ cache IConnection as a property and volative attr?
     """
 
-    key_type_id = 'five.intid.keyreference'
+    key_type_id = "five.intid.keyreference"
     # Default dbname where the root is. This is defined here for
     # backward compatibility with previously created objects.
-    root_dbname = 'main'
+    root_dbname = "main"
 
     def __init__(self, wrapped_obj):
 
         # make sure our object is wrapped by containment only
         try:
-            self.path = '/'.join(wrapped_obj.getPhysicalPath())
+            self.path = "/".join(wrapped_obj.getPhysicalPath())
         except AttributeError:
             self.path = None
 
         # If the path ends with /, it means the object had an empty id.
         # This means it's not yet added to the container, and so we have
         # to defer.
-        if self.path is not None and self.path.endswith('/'):
+        if self.path is not None and self.path.endswith("/"):
             raise NotYet(wrapped_obj)
         self.object = aq_base(wrapped_obj)
         connection = IConnection(wrapped_obj, None)
 
-        if not getattr(self.object, '_p_oid', None):
+        if not getattr(self.object, "_p_oid", None):
             if connection is None:
                 raise NotYet(wrapped_obj)
             connection.add(self.object)
@@ -107,10 +106,10 @@ class KeyReferenceToPersistent(KeyReferenceToPersistent):
         self.dbname = connection.db().database_name
 
     def __setstate__(self, state):
-        for key in ('root_oid', 'oid'):
+        for key in ("root_oid", "oid"):
             value = state.get(key)
-            if isinstance(value, six.text_type):
-                state[key] = value.encode('utf-8')
+            if isinstance(value, str):
+                state[key] = value.encode("utf-8")
         self.__dict__.update(state)
 
     @property
@@ -131,7 +130,7 @@ class KeyReferenceToPersistent(KeyReferenceToPersistent):
         except KeyError:
             # be paranoid and fall back to the complex OFS traverse for (hypothetical)
             # edge cases
-            logger.debug('fall back to OFS traversal for {0}'.format(self.path))
+            logger.debug(f"fall back to OFS traversal for {self.path}")
             obj = self.root.unrestrictedTraverse(self.path, None)
         if obj is None:
             return self.object
@@ -144,7 +143,7 @@ class KeyReferenceToPersistent(KeyReferenceToPersistent):
             if len(site_chain) and isinstance(site_chain[-1], RequestContainer):
                 req = site_chain[-1]
                 new_obj = req
-                # rebuld the chain with the request at the bottom
+                # rebuild the chain with the request at the bottom
                 for item in reversed(chain):
                     new_obj = aq_base(item).__of__(new_obj)
                 obj = new_obj
@@ -155,12 +154,4 @@ class KeyReferenceToPersistent(KeyReferenceToPersistent):
 
     def __hash__(self):
         # XXX Maybe we should consider to use also other fields for the hash
-        return hash((self.dbname,
-                     self.object._p_oid,
-                     ))
-
-    def __cmp__(self, other):
-        # XXX This makes no sense on Python 3
-        if self.key_type_id == other.key_type_id:
-            return cmp((self.dbname, self.oid), (other.dbname, other.oid))
-        return cmp(self.key_type_id, other.key_type_id)
+        return hash((self.dbname, self.object._p_oid))