+++ CHANGES.rst
@@ -14,8 +14,8 @@ Changelog
 Bug fixes:
 
 
-- Fix depecration warnings.
-  Update ``setup.py`` to depend on ``Zope2`` or ``Zope``, dependening on Python version.
+- Fix deprecation warnings.
+  Update ``setup.py`` to depend on ``Zope2`` or ``Zope``, depending on Python version.
   [jensens] (#1)
 - Fix test to work correctly with ``zope.interface >= 5.1``.
   [jensens] (#17)
@@ -102,7 +102,7 @@ Bug fixes:
 
 Fixes:
 
-- Acquisition-unwrap each item in the aq_iter chain, as ``getSite().__parent__`` might return an object aquired from the original context which breaks the parent loop detection.
+- Acquisition-unwrap each item in the aq_iter chain, as ``getSite().__parent__`` might return an object acquired from the original context which breaks the parent loop detection.
   [thet]
 
 - Prevent errors on ``moveIntIdSubscriber`` when the ``IKeyReference`` adapter